# docs/swagger.yaml
openapi: 3.0.0
info:
  title: traQ API
  description: traQ v2 API
  version: 2.6.3
  license:
    name: MIT
    url: 'https://github.com/traPtitech/traQ/blob/master/LICENSE'
  contact:
    name: traP
    url: 'https://github.com/traPtitech/traQ'
servers:
  - url: https://q.trap.jp/api/1.0
    description: production environment
  - url: https://traq-dev.tokyotech.org/api/1.0
    description: staging environment
paths:
...
  /public/icon/{username}:
    get:
      tags:
        - public
      operationId: getPublicUserIcon
      description: 指定したユーザーのアイコン画像を取得します。
      parameters:
        - name: username
          description: 画像を取得するユーザーのユーザー名
          in: path
          required: true
          schema:
            type: string
      security: []
      responses:
        "200":
          description: 正常に取得できました。
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        "404":
          description: 取得できませんでした。指定したユーザーは存在しません。

  # /public/emoji.json

  # /public/emoji.css

  # /public/emoji/{stampID}
...
  /users/{userID}:
    parameters:
      - $ref: "#/components/parameters/userIdInPath"
    get:
      tags:
        - user
      operationId: getUser
      description: ユーザーの詳細を取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: |+
            正常に取得できました。
            ユーザー情報を返します。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: +|
            正常に取得できませんでした。
            指定したユーザーは存在しません。

  # PATCH /users/{userID}

  # /users/{userID}/status

  # /users/{userID}/password
  /users/{userID}/messages:
    parameters:
      - $ref: "#/components/parameters/userIdInPath"
    get:
      tags:
        - message
      operationId: getDirectMessages
      description: DMチャンネルに存在するメッセージを取得します。
      security:
        - traqOAuth2:
            - read
      parameters:
        - $ref: "#/components/parameters/limitInQuery"
        - $ref: "#/components/parameters/offsetInQuery"
        - $ref: "#/components/parameters/sinceInQuery"
        - $ref: "#/components/parameters/untilInQuery"
        - $ref: "#/components/parameters/inclusiveInQuery"
        - $ref: "#/components/parameters/orderInQuery"
      responses:
        "200":
          description: |+
            正常に取得ができました。
            メッセージの配列を返します。
          headers:
            X-TRAQ-MORE:
              $ref: "#/components/headers/X-TRAQ-MORE"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"
        "400":
          description: |+
            変更に失敗しました
            不正なリクエストです
        "404":
          description: |+
            取得に失敗しました。
            指定したユーザーは存在しません。
    post:
      tags:
        - message
      operationId: postDirectMessage
      description: DMチャンネルにメッセージを投稿します。
      parameters:
        - name: embed
          description: メンション・チャンネルリンクを自動埋め込みする場合に1を指定する
          in: query
          required: false
          schema:
            type: integer
      security:
        - traqOAuth2:
            - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessage"
      responses:
        "201":
          description: |+
            投稿に成功しました。
            投稿されたメッセージが返されます。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: +|
            投稿に失敗しました。
            指定したユーザーは存在しません。
...
  /channels/{channelID}/messages:
    parameters:
      - $ref: "#/components/parameters/channelIdInPath"
    get:
      tags:
        - message
      operationId: getMessages
      description: チャンネルに存在するメッセージを取得します。
      security:
        - traqOAuth2:
            - read
      parameters:
        - $ref: "#/components/parameters/limitInQuery"
        - $ref: "#/components/parameters/offsetInQuery"
        - $ref: "#/components/parameters/sinceInQuery"
        - $ref: "#/components/parameters/untilInQuery"
        - $ref: "#/components/parameters/inclusiveInQuery"
        - $ref: "#/components/parameters/orderInQuery"
      responses:
        "200":
          description: |+
            正常に取得ができました。
            メッセージの配列を返します。
          headers:
            X-TRAQ-MORE:
              $ref: "#/components/headers/X-TRAQ-MORE"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"
        "400":
          description: |+
            変更に失敗しました
            不正なリクエストです
        "404":
          description: |+
            取得に失敗しました。
            指定したチャンネルは存在しません。
    post:
      tags:
        - message
      operationId: postMessage
      description: チャンネルにメッセージを投稿します。
      parameters:
        - name: embed
          description: メンション・チャンネルリンクを自動埋め込みする場合に1を指定する
          in: query
          required: false
          schema:
            type: integer
      security:
        - traqOAuth2:
            - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessage"
      responses:
        "201":
          description: |+
            投稿に成功しました。
            投稿されたメッセージが返されます。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: +|
            投稿に失敗しました。
            指定したチャンネルは存在しません。
...
  /messages/{messageID}:
    parameters:
      - $ref: "#/components/parameters/messageIdInPath"
    get:
      tags:
        - message
      operationId: getMessage
      description: 指定したメッセージを取得します。
      security:
        - traqOAuth2:
            - read
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: 取得できませんでした。指定されたメッセージは存在しません。
    put:
      tags:
        - message
      operationId: editMessage
      description: 指定したメッセージを編集します。
      security:
        - traqOAuth2:
            - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessage"
      responses:
        "204":
          description: 正常に編集できました。
        "403":
          description: 編集できませんでした。指定されたメッセージを編集する権限がありません。
        "404":
          description: 編集できませんでした。指定されたメッセージは存在しません。
    delete:
      tags:
        - message
      operationId: deleteMessage
      description: 指定したメッセージを削除します。
      security:
        - traqOAuth2:
            - write
      responses:
        "204":
          description: 正常に削除できました。
        "403":
          description: 削除できませんでした。指定されたメッセージを削除する権限がありません。
        "404":
          description: 削除できませんでした。指定されたメッセージは存在しません。
...
  /messages/{messageID}/stamps:
    parameters:
      - $ref: "#/components/parameters/messageIdInPath"
    get:
      tags:
        - stamp
      operationId: getMessageStamps
      description: 指定したメッセージに押されているスタンプを全て取得します。
      security:
        - traqOAuth2:
            - read
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageStamp"
        "404":
          description: 取得できませんでした。指定したメッセージは存在しません。
...
  /files/{fileID}/meta:
    parameters:
      - $ref: "#/components/parameters/fileIdInPath"
    get:
      tags:
        - file
      operationId: getFileMeta
      description: 指定したファイルのメタデータを取得します
      security:
        - traqOAuth2:
            - read
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileInfo"
        "403":
          description: 取得できませんでした。指定されたファイルにアクセスする権限がありません。
        "404":
          description: 取得できませんでした。指定されたファイルは存在しません。
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: r_session
    traqOAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /api/1.0/oauth2/authorize
          tokenUrl: /api/1.0/oauth2/token
          scopes:
            read: 読み取りスコープ
            write: 書き込みスコープ
            manage_bot: bot関連読み書きスコープ
  parameters:
    channelIdInPath:
      name: channelID
      description: 操作の対象となるチャンネルのID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    userIdInPath:
      name: userID
      description: 操作の対象となるユーザーID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    messageIdInPath:
      name: messageID
      in: path
      required: true
      description: 操作の対象となるメッセージID
      schema:
        type: string
        format: uuid
...
    fileIdInPath:
      name: fileID
      description: 操作の対象となるファイルID
      in: path
      required: true
      schema:
        type: string
...
    limitInQuery:
      in: query
      name: limit
      schema:
        type: integer
        default: 0
      description: 取得する件数 1-200
      example: 50
    offsetInQuery:
      in: query
      name: offset
      schema:
        type: integer
        default: 0
      description: 取得するオフセット
      example: 150
    sinceInQuery:
      in: query
      name: since
      schema:
        type: string
        format: date-time
        default: "0000-01-01T00:00:00.000000Z"
      description: 取得する時間範囲の開始日時
      example: "2016-10-12T11:00:00.000000Z"
    untilInQuery:
      in: query
      name: until
      schema:
        type: string
        format: date-time
      description: 取得する時間範囲の終了日時
      example: "2016-10-12T11:00:00.0000000Z"
    inclusiveInQuery:
      in: query
      name: inclusive
      schema:
        type: boolean
        default: false
      description: 範囲の端を含めるかどうか
    orderInQuery:
      in: query
      name: order
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
      description: 昇順か降順か
  headers:
    X-TRAQ-MORE:
      schema:
        type: boolean
      description: 指定した範囲に要素がさらに存在するかどうか
  schemas:
...
    UserStatus:
      type: integer
      enum:
        - 0
        - 1
        - 2
      description: アカウントの状態 (0:停止,1:有効,2:一時停止)
...
    UserDetail:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: ユーザーUUID
        name:
          type: string
          description: traQ ID
        displayName:
          type: string
          description: 表示名
        iconFileId:
          type: string
          format: uuid
          description: アイコンファイルUUID
        bot:
          type: boolean
          description: BOTかどうか
        twitterId:
          type: string
          description: ツイッターID
        lastOnline:
          type: string
          format: date-time
          description: 最終オンライン日時(オンラインに１度もなってない場合はnull)
        isOnline:
          type: boolean
          description: 現在オンラインかどうか
        suspended:
          type: boolean
          description: アカウントが停止中かどうか
        accountStatus:
          $ref: "#/components/schemas/UserStatus"
        tagList:
          $ref: "#/components/schemas/TagList"
...
    Tag:
      type: object
      properties:
        tagId:
          type: string
          format: uuid
          description: タグUUID
        tag:
          type: string
          description: タグ文字列
        isLocked:
          type: boolean
          description: タグがロックされているかどうか
        createdAt:
          type: string
          format: date-time
          description: タグ付与日時
        updatedAt:
          type: string
          format: date-time
          description: タグ更新日時
    TagList:
      type: array
      items:
        $ref: "#/components/schemas/Tag"
...
    Message:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
          description: メッセージUUID
        userId:
          type: string
          format: uuid
          description: 投稿者UUID
        parentChannelId:
          type: string
          description: 投稿先チャンネルUUID
        content:
          type: string
          description: メッセージ本体
        createdAt:
          type: string
          format: date-time
          description: メッセージ投稿日時
        updatedAt:
          type: string
          format: date-time
          description: メッセージ更新日時
        pin:
          type: boolean
          description: ピン留めされているかどうか
        reported:
          type: boolean
          description: 自分が通報しているかどうか
        stampList:
          type: array
          description: メッセージスタンプ配列
          items:
            $ref: "#/components/schemas/MessageStamp"
    MessageList:
      type: array
      items:
        $ref: "#/components/schemas/Message"
    SendMessage:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Markdown形式のメッセージ本文
          example: Raskって誰？
...
    MessageStamp:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: ユーザーUUID
        stampId:
          type: string
          format: uuid
          description: スタンプUUID
        count:
          type: integer
          description: 押された個数
        createdAt:
          type: string
          format: date-time
          description: 最初に押した日時
        updatedAt:
          type: string
          format: date-time
          description: 最後に押した日時
...
    FileInfo:
      type: object
      properties:
        fileId:
          type: string
          format: uuid
          description: ファイルUUID
        name:
          type: string
          description: ファイル名
        mime:
          type: string
          description: MIMEタイプ
        size:
          type: integer
          description: ファイルサイズ
        md5:
          type: string
          description: MD5ハッシュ
        hasThumb:
          type: boolean
          description: サムネイルがあるかどうか
        thumbWidth:
          type: integer
          description: サムネイル幅
        thumbHeight:
          type: integer
          description: サムネイル高さ
        datetime:
          type: string
          format: date-time
          description: アップロード日時
...
security:
  - cookieAuth: []